backend/requirements.txt
flask
flask-socketio
psycopg2-binary
eventlet
